Now CI unittest will run properly with only one framework version for time limitation (for now).
