+++ Ductus.FluentDocker.MsTest/Ductus.FluentDocker.MsTest.csproj
@@ -27,7 +27,6 @@ Ms Test Support to allow for create, run one or more docker images while testing
 Documentation: https://github.com/mariotoffia/FluentDocker
     </Description>
     <Copyright>© 2016 - 2019 Mario Toffia</Copyright>
-    <DefineConstants Condition=" '$(TargetFramework)' == 'netcoreapp1.1' ">$(DefineConstants);COREFX</DefineConstants>
     <AssemblyOriginatorKeyFile>..\keypair.snk</AssemblyOriginatorKeyFile>
     <SignAssembly>true</SignAssembly>
     <PublicSign Condition=" '$(OS)' != 'Windows_NT' ">true</PublicSign>
